Move SecurityInterceptor so it can be found during Tomcat startup
